+++ src/Request/Channel/OpeningHoursDayRequest.php
@@ -2,7 +2,7 @@
 
 namespace StadGent\Services\OpeningHours\Request\Channel;
 
-use StadGent\Services\OpeningHours\Request\RequestAbstract;
+use DigipolisGent\API\Client\Request\AbstractRequest;
 use StadGent\Services\OpeningHours\Uri\Channel\OpeningHoursDayUri;
 
 /**
@@ -10,7 +10,7 @@ use StadGent\Services\OpeningHours\Uri\Channel\OpeningHoursDayUri;
  *
  * @package StadGent\Services\OpeningHours\Request\Channel
  */
-class OpeningHoursDayRequest extends RequestAbstract
+class OpeningHoursDayRequest extends AbstractRequest
 {
     /**
      * Get the OpeningHours for a single day by the Service & Channel ID.