+++ tests/Service/Service/ServiceServiceGetByOpenDataUriTest.php
@@ -94,7 +94,7 @@ class ServiceServiceGetByOpenDataUriTest extends ServiceTestBase
      *
      * @param \StadGent\Services\OpeningHours\Value\Service $service
      *
-     * @return \StadGent\Services\OpeningHours\Client\ClientInterface
+     * @return \DigipolisGent\API\Client\ClientInterface
      */
     protected function createClientForService(Service $service)
     {