+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All Notable changes to `gent/services-opening-hours` package.
 
+## [1.2.0]
+
+### Changed
+
+* Adapted implementation to use
+  [digipolisgent/api-client](https://github.com/digipolisgent/php_package_dg-api-client).
+
 ## [1.1.0]
 
 ### Changed
@@ -58,9 +65,10 @@ hours for a Channel.
 * DMOH-17: Added the OpeningHoursPeriod & OpeningHoursPeriodHTML methods to the
   ChannelService.
 
+[Unreleased]: https://github.com/StadGent/php_package_services-opening-hours/compare/master...develop
+[1.2.0]: https://github.com/StadGent/php_package_services-opening-hours/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/StadGent/php_package_services-opening-hours/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/StadGent/php_package_services-opening-hours/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/StadGent/php_package_services-opening-hours/compare/0.2.0...1.0.0
 [0.2.0]: https://github.com/StadGent/php_package_services-opening-hours/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/StadGent/php_package_services-opening-hours/releases/tag/0.1.0
-[Unreleased]: https://github.com/StadGent/php_package_services-opening-hours/compare/master...develop